fix: corrigir campo valor estimado para BigDecimal e correção no teste de EntidadeServiceTest

+++ src/main/java/br/gov/es/siscap/dto/ProjetoListaDto.java
@@ -3,14 +3,14 @@ package br.gov.es.siscap.dto;
 import br.gov.es.siscap.models.Microrregiao;
 import br.gov.es.siscap.models.Projeto;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ProjetoListaDto(
         Long id,
         String sigla,
         String titulo,
-        BigInteger valorEstimado,
+        BigDecimal valorEstimado,
         List<String> nomesMicrorregioes) {
 
     public ProjetoListaDto(Projeto projeto) {
